Move get_identity_msisdn to general utils
